Rename and rework the re-compute feature action.
- Properly decoupled from the controller now.
- Offers a static method for clients.
